fix(formula): rename formula file
