+++ packages/engine-formula/src/functions/compatibility/function-map.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { Concatenate } from './concatenate/concatenate';
+import { Concatenate } from './concatenate';
 import { FUNCTION_NAMES_COMPATIBILITY } from './function-names';
 
 export const functionCompatibility = [[Concatenate, FUNCTION_NAMES_COMPATIBILITY.CONCATENATE]];
