+++ packages/engine-formula/src/functions/lookup/function-map.ts
@@ -15,8 +15,8 @@
  */
 
 import { FUNCTION_NAMES_LOOKUP } from './function-names';
-import { Indirect } from './indirect/indirect';
-import { Offset } from './offset/offset';
+import { Indirect } from './indirect';
+import { Offset } from './offset';
 
 export const functionLookup = [
     [Indirect, FUNCTION_NAMES_LOOKUP.INDIRECT],
