+++ packages/engine-formula/src/functions/lookup/offset/index.ts
