+++ packages/engine-formula/src/functions/math/power/index.ts
